requirements.txt
# Core training
torch>=2.4.0
transformers>=4.51.0
trl>=0.15.2
unsloth>=2025.3.0
unsloth_zoo>=2025.3.0
peft>=0.17.0
vllm>=0.8.0
accelerate>=1.6.0
bitsandbytes>=0.45.0

# Style and semantic models
sentence-transformers>=4.0.0
bert-score>=0.3.13

# Data handling
datasets>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# NLP utilities
spacy>=3.7.0
nltk>=3.9.0
sacrebleu>=2.5.0

# Privacy / evaluation
mauve-text>=0.4.0
hdbscan>=0.8.40
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# Misc
wandb>=0.18.0
scipy>=1.11.0
matplotlib>=3.7.0
tqdm>=4.66.0
faiss-gpu>=1.7.2    # or faiss-cpu if no GPU

# Baselines (optional — install only if running baselines)
# pip install presidio-analyzer presidio-anonymizer spacy
# python -m spacy download en_core_web_lg  # for Presidio
